# gestion-app/target/generated-sources/openapi/src/main/java/fr/cours/info/tp/ferrovaire/api/TrainApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package fr.cours.info.tp.ferrovaire.api;

import fr.cours.info.tp.ferrovaire.api.models.Train;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-25T16:23:50.329334900+01:00[Europe/Paris]")
@Validated
@Tag(name = "train", description = "the train API")
public interface TrainApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /train/consulter : Consulter un train par son matricule
     *
     * @param matricule Le matricule unique du train (required)
     * @return Train trouvé avec succès (status code 200)
     *         or Train non trouvé (status code 404)
     */
    @Operation(
        operationId = "consulterTrain",
        summary = "Consulter un train par son matricule",
        responses = {
            @ApiResponse(responseCode = "200", description = "Train trouvé avec succès", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Train.class))
            }),
            @ApiResponse(responseCode = "404", description = "Train non trouvé")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/train/consulter",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Train> consulterTrain(
        @NotNull @Parameter(name = "matricule", description = "Le matricule unique du train", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "matricule", required = true) String matricule
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"matricule\" : \"matricule\", \"vitesseMax\" : 0, \"wagons\" : [ { \"capacite\" : 6, \"id\" : \"id\" }, { \"capacite\" : 6, \"id\" : \"id\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}